CNS/prediction/streamlit_crowd_ui_fixed_clean.py
"""
Complete Streamlit UI for Live Crowd Prediction - Fixed Version
Bhai, yeh working version hai with proper indentation
"""

import streamlit as st
import cv2
import numpy as np
import pandas as pd
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import time
from datetime import datetime, timedelta
import json
from PIL import Image
import requests
from typing import Dict, List, Tuple, Optional

# Import our Central Nervous System
from live_crowd_predictor import (
    central_nervous_system, 
    get_central_status, 
    start_central_monitoring, 
    stop_central_monitoring,
    get_live_crowd_status,  # Legacy support
    start_live_monitoring,  # Legacy support
    stop_live_monitoring    # Legacy support
)

# ... existing code ...
# (The rest of the file is copied exactly as in streamlit_crowd_ui_fixed.py, with consistent 4-space indentation) 