CNS/prediction/requirements-basic.txt
# Core dependencies
numpy==1.24.3
geohash2==1.1
python-dotenv==1.0.0

# GCP libraries
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0

# Computer Vision
opencv-python==4.8.1.78
pillow==10.0.1

# Data processing
pandas==2.1.3